refactor: controller를 하나로 통합
로직을 한 곳에 모으기 위해 controller를 하나로 통합
