+++ src/main/java/vendingmachine/Application.java
@@ -1,13 +1,11 @@
 package vendingmachine;
 
-import vendingmachine.controller.CoinController;
-import vendingmachine.controller.ProductController;
+import vendingmachine.controller.VendingMachineController;
 
 public class Application {
 	public static void main(String[] args) {
-		CoinController coinController = new CoinController();
-		ProductController productController = new ProductController();
-		coinController.setVendingMachineCoinCounter();
-		productController.setProducts();
+		VendingMachineController vendingMachineController = new VendingMachineController();
+		vendingMachineController.setVendingMachineCoinCounter();
+		vendingMachineController.setProducts();
 	}
 }
